hf-backend-deploy/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
sqlmodel>=0.0.22
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
asyncpg>=0.30.0
psycopg2-binary>=2.9.9
pydantic>=2.10.5
pydantic-settings>=2.7.1
python-dotenv>=1.0.0
email-validator>=2.0.0
python-dateutil>=2.8.2
bcrypt>=4.0.0
